Moved GUISound initialization and construction since it can be done safely outside of AudioMan now Added comment to SoundContainer lambda, fixed non-referenced based range for loop, added constnesses to GetSoundDataForSound method. Fixed Vector to FMOD_VECTOR conversion trying to use scene dimensions if the scene isn't loaded, since the assert makes this unsafe. Fixed bad typing on sound max audible distance that would cut it off (it was a u short instead of a u int) Removed junk member variable from GUISound (poopsound)

+++ System/Constants.h
@@ -42,7 +42,7 @@ namespace RTE {
 	static constexpr unsigned short c_MinimumDistanceForPanning = 50; //!< The minimum distance before which sound panning should not occur. Not relevant for immobile sounds or in splitscreen.
 	*/
 	static constexpr unsigned short c_DefaultAttenuationStartDistance = 100; //!< The default start distance for attenuating sounds. Individual sounds can have different values for this.
-	static constexpr unsigned short c_SoundMaxAudibleDistance = 100000; //!< The maximum distance at which any sound can possibly be heard, after which point it will have 0 volume. Arbitrary default suggested by FMOD.
+	static constexpr unsigned int c_SoundMaxAudibleDistance = 100000; //!< The maximum distance at which any sound can possibly be heard, after which point it will have 0 volume. Arbitrary default suggested by FMOD.
 #pragma endregion
 
 #pragma region Network Constants
